backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asyncpg==0.29.0
asyncssh==2.17.0
attrs==24.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
bs4==0.0.2
bson==0.5.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
cryptography==43.0.1
decorator==5.1.1
Deprecated==1.2.14
deptry==0.20.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fabric==3.2.2
fastapi==0.115.2
fastapi-utilities==0.3.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
frozenlist==1.4.1
gevent==24.10.3
geventhttpclient==2.3.1
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.4.0
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.32.0
loguru==0.7.2
Mako==1.3.5
MarkupSafe==3.0.1
msgpack==1.1.0
multidict==6.1.0
opentelemetry-api==1.27.0
opentelemetry-distro==0.48b0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-logging==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
paramiko==3.5.0
passlib==1.7.4
prometheus_client==0.21.0
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-logging-loki==0.3.1
pytz==2024.2
pyzmq==26.2.0
redis==5.1.1
requests==2.32.3
rfc3339==6.2
rsa==4.9
sentry-sdk==2.17.0
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.39.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
Werkzeug==3.0.4
win32-setctime==1.1.0
wrapt==1.16.0
yarl==1.15.2
zipp==3.20.2
zope.event==5.0
zope.interface==7.1.0